requirements.txt
telethon
psycopg2-binary
pgvector
fastapi
uvicorn[standard]
sqlalchemy
openai
aiogram
alembic
python-dotenv
httpx
Pillow
pytest
pytest-asyncio
python-multipart
pre-commit
black
flake8
